Get two-way listener working
Redo of the two-way mode using the .net api
Wait for 250ms before actually firing the updates, as that means we've unpacked the entire callback.
Move the data produced by two way to the nodes _studio and _canvas signifying where the data comes from.
Added the ability to reset the model with a reload and force update of the external sources.
After updating all the tunables are deleted.
Additional runtime setting to update all external sources and not just the one caller. This still requires path and secret from the command line.

To work properly, the data should be sent [makes everything up to date], a state saved and then the data should be re-sent.  This makes sure the correct state file is given to firebase.