Refactor Listener
Refactor the whole app to run in multiple threads. Each of the components has it's own thread,:
- StudioListener manages the underlying components and sends the heartbeats.
- StudioReceiver listens on the `_studio/callback` key for any data, which is passed to ModelCaller
- ModelCaller collates the updates and calls the API to actually update the model, results are sent to StudioSender
- StudioSender updates the Studio firebase database
The application nows supports an additional argument to set the logging level.
Logging can also be changed at runtime by pressing the number keys 0 to 9; 0==errors & 9==debugs. Default is 3, 4/5 is good for live server. 9 is a lot of noise.